backend/pyproject.toml
[project]
name = "pihole-wtm"
version = "0.1.0"
description = "Pi-hole dashboard enriched with WhoTracksMe/Ghostery tracker intelligence"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",  # used for TrackerDB lookups and local query database
    "cachetools>=5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-cachetools",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S"]

[tool.ruff.lint.per-file-ignores]
# pytest uses assert statements by design; long descriptive test names and hardcoded
# test credentials (S105) are intentional in test files
"tests/**/*.py" = ["S101", "S105", "S106", "E501"]
# Dynamic SQL uses f-strings for structure only; all user input goes through parameterised queries
"app/services/database.py" = ["S608"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
